Mise a jour des dtos du service de gestion des documents

+++ src/users/users.controller.ts
@@ -9,52 +9,88 @@ import {
   UseGuards,
   HttpCode,
   HttpStatus,
+  NotFoundException,
+  ForbiddenException,
 } from '@nestjs/common';
 import { UsersService } from './users.service';
 import { CreateUserDto } from '../../libs/common/src/dto/user/create-user.dto';
 import { UpdateUserDto } from '../../libs/common/src/dto/user/update-user.dto';
-import { AccessTokenGuard, RolesGuard } from 'libs/common/src/guards';
-import { Role } from './entities/user.entity';
-import { Roles } from 'libs/common/src/decorators';
+import { AccessTokenGuard } from 'libs/common/src/guards';
+import { Role, User } from './entities/user.entity';
+import { GetCurrentUser} from 'libs/common/src/decorators';
+import { Actions } from 'libs/common/src/enums/actions.enum';
+import { CreatePasswordDto } from 'libs/common/src/dto/user/create-password.dto';
+import { Action } from 'rxjs/internal/scheduler/Action';
+
 
 @Controller('users')
-@UseGuards(RolesGuard)
-@UseGuards(AccessTokenGuard)
+
 export class UsersController {
   constructor(private readonly usersService: UsersService) {}
 
   @Post()
-  @Roles((Role.Admin))
   @HttpCode(HttpStatus.CREATED)
-  create(@Body() createUserDto: CreateUserDto) {
+  async create(@Body() createUserDto: CreateUserDto, @GetCurrentUser() currentUser: User) {
     return this.usersService.create(createUserDto);
   }
 
   @Get()
-  @Roles((Role.Admin))
   @HttpCode(HttpStatus.OK)
-  findAll() {
-    return this.usersService.findAll();
+  async findAll(@GetCurrentUser() currentUser: User) {
+    return await this.usersService.findAll();
   }
 
   @Get(':id')
-  @Roles((Role.Admin))
   @HttpCode(HttpStatus.OK)
-  findOne(@Param('id') id: string) {
-    return this.usersService.findOne(+id);
+  async findOne(@Param('id') id: string, @GetCurrentUser() currentUser: User) {
+    const userResearch: User = await this.usersService.findOne(+id);
+    if (!userResearch) {
+      throw new NotFoundException(`
+        Cet utilisateur n'existe pas. 
+      `)
+    }
+    return userResearch;
   }
 
   @Patch(':id')
-  @Roles((Role.Admin))
   @HttpCode(HttpStatus.OK)
-  update(@Param('id') id: string, @Body() updateUserDto: UpdateUserDto) {
-    return this.usersService.update(+id, updateUserDto);
+  async update(@Param('id') id: string, 
+         @Body() updateUserDto: UpdateUserDto, 
+         @GetCurrentUser() currentUser: User) {
+    const userResearch: User = await this.usersService.findOne(+id)  
+      return await this.usersService.update(+id, updateUserDto);
+  }
+
+  @Patch('/password/:id')
+  @HttpCode(HttpStatus.OK)
+  async updatePassword(@Param('id') id: string, 
+                       @Body() createPasswordDto: CreatePasswordDto,
+                       @GetCurrentUser() currentUser: User) : Promise<User> {
+    return await this.usersService.updatePassword(createPasswordDto.password, +id)
+  }
+
+  @Post('/password/:token')
+  @HttpCode(HttpStatus.CREATED)
+  async createPassword(@Param('token') token: string, 
+                       @Body() createPasswordDto: CreatePasswordDto) : Promise<User> {
+    const userResearch: User = await this.usersService.findUserByCreatePasswordToken(token);
+    if (!userResearch) {
+      throw new NotFoundException(`
+        L'utilisateur n'existe pas
+      `)
+    }
+    return await this.usersService.createPassword(createPasswordDto.password, userResearch);
   }
 
   @Delete(':id')
-  @Roles((Role.Admin))
   @HttpCode(HttpStatus.OK)
-  remove(@Param('id') id: string) {
+  async remove(@Param('id') id: string, @GetCurrentUser() currentUser: User) {
+    const userResarch: User = await this.usersService.findOne(+id);
+    if (!userResarch) {
+      throw new NotFoundException(`
+        Cet utilisateur d'existe pas
+      `)
+    }
     return this.usersService.remove(+id);
   }
 }
